veracrpytFINAL/requirements.txt
# requirements.txt
streamlit==1.32.0
pymongo==4.6.1
python-dotenv==1.0.0
cryptography==41.0.7
dnspython==2.4.2
